+++ gym_space_engineers/__init__.py
@@ -23,6 +23,28 @@ register(
 )
 
 register(
+    id="SE-WalkingTest-v1",
+    entry_point="gym_space_engineers.envs:WalkingRobotIKEnv",
+    # max_episode_steps=200,  # around 20s of interaction
+    kwargs={
+        "control_frequency": 10,  # 10Hz
+        "symmetric_control": True,
+        "add_end_effector_velocity": True,
+    },
+)
+
+register(
+    id="SE-Forward-v1",
+    entry_point="gym_space_engineers.envs:WalkingRobotIKEnv",
+    max_episode_steps=200,  # around 20s of interaction
+    kwargs={
+        "control_frequency": 10,  # 10Hz
+        "symmetric_control": False,
+        "add_end_effector_velocity": True,
+    },
+)
+
+register(
     id="SE-Symmetric-v1",
     entry_point="gym_space_engineers.envs:WalkingRobotIKEnv",
     max_episode_steps=200,  # around 20s of interaction
@@ -39,11 +61,43 @@ register(
     max_episode_steps=200,  # around 20s of interaction
     kwargs={
         "control_frequency": 10,  # 10Hz
-        "symmetric_control": True,
+        "symmetric_control": False,
         "add_end_effector_velocity": True,
         "task": "turn_left",
     },
 )
+
+register(
+    id="SE-Generic-v1",
+    entry_point="gym_space_engineers.envs:WalkingRobotIKEnv",
+    max_episode_steps=200,  # around 20s of interaction
+    kwargs={
+        "control_frequency": 10,  # 10Hz
+        "symmetric_control": False,
+        "add_end_effector_velocity": True,
+        "task": "generic_locomotion",
+        "weight_distance_traveled": 1.0,
+        "weight_heading_deviation": 20.0,
+        "desired_angular_speed": 10.0,
+        "desired_linear_speed": 4.0,
+    },
+)
+
+
+register(
+    id="SE-MultiTask-v1",
+    entry_point="gym_space_engineers.envs:WalkingRobotIKEnv",
+    max_episode_steps=200,  # around 20s of interaction
+    kwargs={
+        "control_frequency": 10,  # 10Hz
+        "symmetric_control": False,
+        "add_end_effector_velocity": True,
+        "task": "forward",
+        "randomize_task": True,
+        "randomize_interval": 100,  # change task every 10s
+    },
+)
+
 # No Timeout
 register(
     id="SpaceEngineers-WalkingRobot-IK-v1",